requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
kafka-python==1.3.4
MarkupSafe==1.0
Werkzeug==0.12.2
